+++ setup.py
@@ -33,9 +33,6 @@ with open(README_PATH) as readme_file:
     )
 
 
-csplayout_requirements = [
-    "python-constraint>=1.4",
-]
 visualization_extras = [
     "matplotlib>=3.3",
     "ipywidgets>=7.3.0",
@@ -49,6 +46,7 @@ z3_requirements = [
     "z3-solver>=4.7",
 ]
 bip_requirements = ["cplex", "docplex"]
+csp_requirements = ["python-constraint>=1.4"]
 
 
 setup(
@@ -85,11 +83,11 @@ setup(
         "visualization": visualization_extras,
         "bip-mapper": bip_requirements,
         "crosstalk-pass": z3_requirements,
-        "csp-layout-pass": csplayout_requirements,
+        "csp-layout-pass": csp_requirements,
         # Note: 'all' does not include 'bip-mapper' because cplex is too fiddly and too little
         # supported on various Python versions and OSes compared to Terra.  You have to ask for it
         # explicitly.
-        "all": visualization_extras + z3_requirements + csplayout_requirements,
+        "all": visualization_extras + z3_requirements + csp_requirements,
     },
     project_urls={
         "Bug Tracker": "https://github.com/Qiskit/qiskit-terra/issues",
